+++ src/service/drone/impl/drone_service.cpp
@@ -194,7 +194,7 @@ void DroneService::write_back_pdu()
     ServicePduDataType actuator_pdu_data = {};
     actuator_pdu_data.id = SERVICE_PDU_DATA_ID_TYPE_ACTUATOR_CONTROLS;
     for (int i = 0; i < ROTOR_NUM; i++) {
-        actuator_pdu_data.pdu.actuator_controls.controls[i] = pwm_duty_.d[i];
+        actuator_pdu_data.pdu.actuator_controls.controls[i] = static_cast<float>(pwm_duty_.d[i]);
     }
     write_pdu(actuator_pdu_data);
